refactors TypeRelations component
